fix: 修复并发安全和Java 8兼容性问题，完善文档注释
主要变更:
- BloomFilter: 修复负数hash问题(Math.abs MIN_VALUE)，添加边界检查
- LSMTree: 添加memTableLock减少锁竞争，使用AtomicLong防止文件名冲突
- 统一WAL和MemTable的timestamp确保数据一致性
- 修复Java 8兼容性问题(ByteBuffer.flip, Math.floorMod)
- 移除GenericObjectPool后台回收线程避免资源泄漏
- 完善所有核心类的Javadoc文档注释
- 新增9个测试用例验证修复和边界场景
- build.sh: 确保日志目录存在

+++ src/main/java/com/brianxiadong/lsmtree/ConsistentHashPartitionStrategy.java
@@ -6,11 +6,21 @@ import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * 一致性哈希分区策略
+ * <p>
+ * 使用 SHA-256 哈希值对分区数取模来决定 Key 所属的分区。
+ * 注意：这里简化为模运算，标准的一致性哈希通常使用哈希环（Hash Ring）来支持动态扩缩容。
+ * <p>
+ * 这种策略可以均匀分布数据，但不支持高效的范围查询（必须扫描所有分区）。
+ */
 public class ConsistentHashPartitionStrategy implements PartitionStrategy {
     @Override
     public int getPartition(String key, int numPartitions) {
         long h = hash(key);
-        long idx = Math.floorMod(h, numPartitions);
+        // 兼容 Java 8: Math.floorMod(long, int) 是 Java 9 引入的
+        // 使用 Math.floorMod(long, long)
+        long idx = Math.floorMod(h, (long) numPartitions);
         return (int) idx;
     }
